pyproject.toml
[tool.poetry]
name = "giuseppe"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pillow = "^8.1.2"
numpy = "^1.20.1"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
ipdb = "^0.13.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
